SWESARR Tutorial Update
Updated SWESARR Tutorial. Includes comparison to SWE values provided by SnowExSQL and a few extra visualizations for comparing SWE and SWESARR data.

The default RAM allocation (2GB) for CryoInTheCloud can't handle the final figure generated by this notebook. This is poor memory management on my part. I've commented out the figure for the tutorial. Currently, this notebook uses 1GB of RAM.

Most libraries are contained by default via CryoInTheCloud, but SnowExSQL isn't. This may be an issue with merging to the main branch.